Add files via upload
Main.java
Moved around code in main and set primaryStage to be fullscreen. 
primaryStage and the gameScene now have a size that is the height and width of a person’s monitor.
The canvas size is now based on the height and width of our camera object in Game.

Game.java
Removed top and bottom node from border layout, our Border Layout now only contains a Canvas in the center. We can implement an information pane for our game later one.
Moved the camera.centerOnCharacter method to the Chracter class file, it will now center on character whenever it actually moves instead of every tick.
Changed .setOnKeyPressed and .setOnKeyRleased method in game to just assign them to the gameScene directly instead of creating a new scene object.
Added an ArrayList of enemies and created 5 enemies in the constructor. The game.tick() method will run through a for loop and render these enemies.

Camera.java
Just cleaned up the code a bit and removed useless variables, added some commenting to explain what it is used and needed for.
Main thing that fixed our camera and lets it properly follow that character was that we had to set the maxOffset = mapWidth – viewWidth.

Enemy.java
Create default values for enemy including: Health, Strength, Defense, Width, Height.
We can probably do the same thing for character, so we don’t have to use crazy long constructors.
Can now use a constructor that takes an x and y position, using int values. This position will be based on pixel values, if want the enemy on a specific square just multiply by Tile.TILEWIDTH / TILEHEIGHT.

Tile.java
Just changed the default tile width/height to 90.